Fix Issue #410: worksheet protection settings are not retained by openxlsx; Add protectWorksheet function to enable worksheet-level protection (potentially password-protected)
This patch loads the <sheetProtection> XML tag from the sheet's xml file, stores it in the corresponding field of the Worksheet and simply writes it out again on saveWorkbook.

To modify the worksheet-level protection, the protectWorksheet function is provided. It can either replace the protection or remove it. It will not modify the flags of the existing worksheet protection.