Improve handling of non-region names in getNamedRegions and add related test
Excel names don't need to point to regions, they may point to constants, such as 10000, or other names.
Currently getNamedRegions can skip these, causing the sheets and regions attributes not to line up with the names.
This patch just sets the sheet and region to "" (empty string) for the names like this.